Use double-precision to store split thresholds (#55)

+++ include/treelite/base.h
@@ -14,7 +14,7 @@
 namespace treelite {
 
 /*! \brief float type to be used internally */
-typedef float tl_float;
+typedef double tl_float;
 /*! \brief feature split type */
 enum class SplitFeatureType : int8_t {
   kNone, kNumerical, kCategorical
